% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-doxygen.R
\name{add_doxygen_to_roclets}
\alias{add_doxygen_to_roclets}
\alias{rm_doxygen_from_roclets}
\title{Add/remove doxygen to default roclet}
\usage{
add_doxygen_to_roclets(base_path = ".")

rm_doxygen_from_roclets(base_path = ".")
}
\arguments{
\item{base_path}{Folder of the package}
}
\value{
The new value of Roxygen options
}
\description{
Add/remove the doxygen roclet from the list of 
    roclets ran by default for this package.
    This is done by modifying the options of the Roxygen package 
    in the DESCRIPTION file
}
\examples{
\dontrun{
add_doxygen_to_roclets()
}
}

