\name{assign.parent}
\alias{assign.parent}
\title{Assign a variable in the parent environment when <<-...}
\usage{assign.parent(var, value, env)}
\description{Assign a variable in the parent environment when \code{<<-}
doesn't see to work.}
\value{NULL}
\arguments{\item{var}{string of the variable to assign}
\item{value}{value to be assigned}
\item{env}{environment of the assignment (\code{environment()})}}
