\name{cat.description}
\alias{cat.description}
\title{Print the field-value pair to a given file or standard out.}
\usage{cat.description(field, value, file="")}
\description{Print the field-value pair to a given file or standard out.}
\value{\code{NULL}}
\arguments{\item{field}{the field to be printed}
\item{value}{the value to be printed}
\item{file}{the file whither to print (a blank string being
standard out)}}
