\name{citation_roclet}
\alias{citation_roclet}
\title{Roclet: make inst/CITATION file.}
\usage{
  citation_roclet()
}
\value{
  citation roclet
}
\description{
  This roclet creates the inst/CITATION file based on a
  inst/CITATION.bib file.
}
\seealso{
  Other roclets: \code{\link{collate_roclet}},
  \code{\link{demo_roclet}},
  \code{\link{namespace_roclet}}, \code{\link{rd_roclet}}
}

