\name{Compose}
\alias{Compose}
\title{Compose an arbitrary number of functions.}
\usage{Compose(...)}
\description{Compose an arbitrary number of functions.
My Happy Hacking keyboard gave out during the writing of this
procedure; moment of silence, please.}
\value{A composed function}
\arguments{\item{\dots}{the functions to be composed}}
