\name{Curry}
\alias{Curry}
\title{Pre-specify a procedures named parameters, returning a new procedure.}
\usage{Curry(FUN, ...)}
\description{Pre-specify a procedures named parameters, returning a new procedure.}
\details{Thanks, Byron Ellis.
\url{https://stat.ethz.ch/pipermail/r-devel/2007-November/047318.html}}
\value{A new function partially determined}
\arguments{\item{FUN}{the function to be curried}
\item{\dots}{the determining parameters}}
