% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-data-format.R
\name{default_data_format}
\alias{default_data_format}
\title{Default format for data}
\usage{
default_data_format(x)
}
\arguments{
\item{x}{A data object}
}
\value{
An object of class \code{rd}, or \code{NULL}.

}
\description{
This function is called to generate the default "Format" section for each
data object.  The default implementation will print the class and dimension
information.
}

