\name{description.dependencies}
\alias{description.dependencies}
\title{Gather a DESCRIPTION's dependencies from the...}
\usage{description.dependencies(description.file)}
\description{Gather a \file{DESCRIPTION}'s dependencies from the
\code{Package}, \code{Depends}, \code{Imports}, \code{Suggests},
and \code{Enhances} fields.}
\value{A list of dependencies}
\section{TODO}{Test this!}
\arguments{\item{description.file}{the \file{DESCRIPTION} to parse}}
