% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-escaping.R
\name{double_escape_md}
\alias{double_escape_md}
\title{Check markdown escaping}
\description{
This is a regression test for Markdown escaping.

Each of the following bullets should look the same when rendered:
\itemize{
\item Backticks: \verb{\\}, \verb{\\\%}, \verb{\\$}, \verb{\\_}
\item \verb{\verb{}}: \verb{\\}, \verb{\\\%}, \verb{\$}, \verb{\_}
}

[ this isn't a link ]

\[ neither is this \]
}
\examples{
"\%" # percent
"\"" # double quote
'\'' # single quote
}
\keyword{internal}
