% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-doxygen.R
\name{doxygen_init}
\alias{doxygen_init}
\title{Prepares the R package structure for use with doxygen}
\usage{
doxygen_init(doxy_file)
}
\arguments{
\item{rootFolder}{The root of the R package}
}
\value{
A boolean: did everything go smoothly? 
}
\description{
Makes a configuration file in inst/doxygen
    and sets a few options:
    \itemize{
       \item{EXTRACT <- ALL = YES}
       \item{INPUT = src/}
       \item{OUTPUT <- DIRECTORY = inst/doxygen/}
    }
}
\note{
Please note this implies the installation of doxygen on the system,
      accessible on the command line by `doxygen`
}
\examples{
\dontrun{
doxygen_init()
}
}
\keyword{internal}

