% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-doxygen.R
\name{doxygen_roclet}
\alias{doxygen_roclet}
\title{Makes doxygen documentation}
\usage{
doxygen_roclet()
}
\description{
Roclet making documentation from
   doxygen instructions in src/, note that a doxygen configuration file
   inst/doxygen/Doxyfile is created. You can edit it but edits would be lost
   if you call this function again or call to the clean method of this roclet. 
}
\details{
The function \code{roc_output.doxygen} verifies the existence of a doxygen structure, if
   needed it creates it and run doxygen. 
   doxygen should be previously installed
   and accessible in the command line of the system as `doxygen`
}
\examples{
\dontrun{
doxygen_roclet()
}
}

