\name{expression.from.partitum}
\alias{expression.from.partitum}
\title{Extract the expression from the parse tree.}
\usage{expression.from.partitum(partitum)}

\description{
  Extract the expression from the parse tree.
}
\value{the extracted expression}
\arguments{
  \item{partitum}{the parsed elements}
}
