% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{generate_template}
\alias{generate_template}
\title{Generate a template params list}
\usage{
generate_template(x, verbose = TRUE)
}
\arguments{
\item{x}{name of the function for which you want a param template}

\item{verbose}{whether to print the template. Takes default from \code{getOption("verbose")}}
}
\description{
This function generates a template of \code{roxygen} documentation for a function. \cr
Also, the \code{example} section of the output is experimental and needs more work \cr
}
\examples{
out <- generate_template(x = "plot", verbose = TRUE)
}

