\name{is.function.definition}
\alias{is.function.definition}
\title{Whether the expression assigns function...}
\usage{is.function.definition(expression)}
\description{Whether the expression assigns function}
\value{Whether the expression assigns a function}
\arguments{\item{expression}{the expression to check for assignment}}
