% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{load_options}
\alias{load_options}
\title{Load roxygen2 options}
\usage{
load_options(base_path = ".")
}
\arguments{
\item{base_path}{Path to package.}
}
\description{
Options can be stored in the \code{Roxygen} field of the \code{DESCRIPTION}, or
in \code{man/roxygen/meta.R}. In either case, the code is parsed and evaluated
in a child of the base environment.
}
\details{
Options in \code{man/roxygen/meta.R} override those present \code{DESCRIPTION}.
}
\keyword{internal}
