% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown_pass1}
\alias{markdown_pass1}
\title{Expand the embedded inline code}
\usage{
markdown_pass1(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
Text with the inline code expanded. A character vector of the
same length as the input \code{text}.
}
\description{
For example this becomes two: 2.
Variables can be set and then reused, within the same
tag:
The value of \code{x} is 100.
}
\keyword{internal}
