\name{maybe.quote}
\alias{maybe.quote}
\title{Quote characters and return all other types untouched.}
\usage{maybe.quote(x)}

\description{
  Quote characters and return all other types untouched.
}
\value{The quoted character or the untouched value}
\arguments{
  \item{x}{the value to maybe quote}
}
