% Generated by roxygen2 (4.0.0): do not edit by hand
\name{md2rd_roclet}
\alias{md2rd_roclet}
\title{Roclet: make Rd files from Markdown-Roxygen comments.}
\usage{
md2rd_roclet()
}
\description{
Supports a limited subset of Markdown syntax in Roxygen comments, see Details.
}
\details{
Supported Markdown syntax: \itemize{ \item itemized lists (lines starting
with a star followed by a blank, subsequent lines for the same item starting
with 4 spaces) \item enumerated lists (first item starting with 1 followed by
a period, subsequent items starting with numbers+period or star, or 4 spaces
to continue same item after a linebreak)\item \code{code}-formatting with
backticks (\code{`<code>`}) \item strong/bold (\code{**<bold>**}) \item
italic/emphasized text (\code{*<emph>*}) \item links to other function docs
(\code{![<function>](<some_package>)} or \code{![function]}) \item inline
math (\code{$<math>$}) \item displayed equations (\code{$$<math>$$})}. Note
that nested lists are not supported, and neither are linebreaks within math
environments or code snippets.
}
\seealso{
\code{tests/test-md2rd.R} for syntax examples.

Other roclets: \code{\link{S3method}},
  \code{\link{export}}, \code{\link{exportClass}},
  \code{\link{exportMethod}}, \code{\link{import}},
  \code{\link{importClassesFrom}},
  \code{\link{importFrom}},
  \code{\link{importMethodsFrom}},
  \code{\link{namespace_roclet}}; \code{\link{rd_roclet}}
}

