\name{name}
\title{Rd value parsers...}
\description{Rd value parsers}
\alias{name}
\alias{aliases}
\alias{title}
\alias{usage}
\alias{references}
\alias{concept}
\alias{note}
\alias{seealso}
\alias{example}
\alias{examples}
\alias{keywords}
\alias{return}
\alias{author}
\alias{TODO}
\seealso{make.Rd.roclet}
