% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-format.R
\name{object_format}
\alias{object_format}
\title{Default format for data}
\usage{
object_format(x)
}
\arguments{
\item{x}{A data object}
}
\value{
A \code{character} value with valid \code{Rd} syntax, or \code{NULL}.
}
\description{
This function is called to generate the default "Format" section for each
data object. The default implementation will return the class and dimension
information.
}
\seealso{
Other extending: 
\code{\link{is_s3_generic}()},
\code{\link{load_options}()},
\code{\link{object}()},
\code{\link{parse_package}()},
\code{\link{rd_section}()},
\code{\link{roc_proc_text}()},
\code{\link{roclet_find}()},
\code{\link{roxy_block}()},
\code{\link{roxy_tag}()},
\code{\link{roxy_tag_rd}()},
\code{\link{tag_parsers}},
\code{\link{tags_list}()}
}
\concept{extending}
