\name{pairwise}
\alias{pairwise}
\title{Combine a list into pairwise elements; lists should...}
\usage{pairwise(list)}
\description{Combine a list into pairwise elements; lists should
be of the same length. In case of odd numbers of members,
the last will be removed.}
\value{A list of pairwise elements}
\arguments{\item{list}{the list to be pairwise decomposed}}
