\name{parse.description.file}
\alias{parse.description.file}
\title{Convenience function to call...}
\usage{parse.description.file(description.file)}
\description{Convenience function to call
\code{\link{parse.description.text}}
with the given \file{DESCRIPTION} file.}
\value{\code{NULL}}
\arguments{\item{description.file}{the \file{DESCRIPTION} file to be parsed}}
