\name{parse.description.text}
\alias{parse.description.text}
\title{Parse lines of text corresponding to a package DESCRIPTION file.}
\usage{parse.description.text(description)}
\description{Parse lines of text corresponding to a package DESCRIPTION file.}
\value{A list of values indexed by field}
\arguments{\item{description}{the lines of tex}}
