\name{parse.element}
\alias{parse.element}
\title{Parse a raw string containing key and expressions.}
\usage{parse.element(element)}
\description{Parse a raw string containing key and expressions.}
\value{A list containing the parsed constituents}
\arguments{\item{element}{the string containing key and expressions}}
