\name{parse.files}
\alias{parse.files}
\title{Parse many files at once.}
\usage{
  parse.files(paths)
}
\arguments{
  \item{\dots}{files to be parsed}
}
\value{
  List containing parsed directives
}
\description{
  Parse many files at once.
}
\seealso{
  \code{\link{parse.file}}
}
\keyword{internal}

