\name{parse.formals}
\alias{parse.formals}
\title{Find the formal arguments associated with a given...}
\usage{parse.formals(expressions)}
\description{Find the formal arguments associated with a given
expression (may be \code{NULL}).}
\value{The formal arguments of said expression or
\code{NULL}}
\arguments{\item{expressions}{the expressions from which to extract
formal arguments}}
