\name{parse.name.description}
\alias{parse.name.description}
\title{Parse an element containing a mandatory name...}
\usage{
  parse.name.description(key, rest)
}
\arguments{
  \item{key}{the parsing key}

  \item{rest}{the expression to be parsed}
}
\value{
  A list containing the key, name and description
}
\description{
  Parse an element containing a mandatory name and
  description (such as \code{@param}).
}
\keyword{internal}

