\name{parse.preref}
\alias{parse.preref}
\title{Resorts to the default parser but with a warning about the...}
\usage{parse.preref(key, rest)}
\description{Resorts to the default parser but with a warning about the
unknown key.}
\value{A list containing the key and expression (possibly
null)}
\seealso{\code{\link{parse.default}}}
\arguments{\item{key}{the parsing key}
\item{rest}{the expression to be parsed}}
