\name{parse.ref.list}
\alias{parse.ref.list}
\title{Parse a preref/srcrefs pair...}
\usage{\method{parse.ref}{list} (ref, ...)}
\arguments{
  \item{ref}{the preref/srcref pair}
  \item{\dots}{ignored}
}
\description{
  Parse a preref/srcrefs pair
}
\value{List combining the parsed preref/srcref}
\keyword{internal}

