\name{parse.ref.preref}
\alias{parse.ref.preref}
\title{Parse a preref...}
\usage{
  \method{parse.ref}{preref} (ref, ...)
}
\arguments{
  \item{ref}{the preref to be parsed}

  \item{\dots}{ignored}
}
\value{
  List containing the parsed preref
}
\description{
  Parse a preref
}
\keyword{internal}

