\name{parse.ref}
\alias{parse.ref}
\title{Parse either srcrefs, prerefs or pairs of the same.}
\usage{
  parse.ref(ref)
}
\arguments{
  \item{ref}{the srcref, preref or pair of the same}
}
\value{
  List containing the parsed srcref/preref
}
\description{
  Parse either srcrefs, prerefs or pairs of the same.
}
\keyword{internal}

