\name{parse.srcref}
\alias{parse.srcref}
\title{By default, srcrefs are ignored; this parser returns list().}
\usage{parse.srcref(pivot, expression)}

\description{
  By default, srcrefs are ignored; this parser returns
  \code{list()}.
}
\value{\code{list()}}
\arguments{
  \item{pivot}{the parsing pivot}
  \item{expression}{the expression to be parsed}
}

