\name{parse.text}
\alias{parse.text}
\title{Text-parsing hack using tempfiles for more facility.}
\usage{parse.text(text)}
\arguments{
  \item{text}{stringr containing text to be parsed}
}
\description{
  Text-parsing hack using tempfiles for more facility.
}
\value{The parse tree}
\keyword{internal}

