\name{parse.toggle}
\alias{parse.toggle}
\title{Turn a binary element on; parameters are ignored.}
\usage{
  parse.toggle(key, rest)
}
\arguments{
  \item{key}{parsing key}

  \item{rest}{the expression to be parsed}
}
\value{
  A list with the key and \code{TRUE}
}
\description{
  Turn a binary element on; parameters are ignored.
}
\keyword{internal}

