% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_code}
\alias{parse_code}
\title{Parse roxygen2 comments in a piece of code.}
\usage{
parse_code(file, text = NULL, env, registry = default_tags(),
  wrap = FALSE, markdown = markdown_global_default,
  file_encoding = "UTF-8")
}
\arguments{
\item{file}{Path of the file to process.}

\item{text}{Instead of specifying \code{file}, users can also specify a character
vector \code{text}, containing the R code.}

\item{env}{An environment containing the result of parsing and evaluating
the file.}

\item{registry}{A roclet tag registry.}

\item{wrap}{Logical; should roxygen2 output be wrapped? \code{FALSE} by default.}

\item{markdown}{Logical value indicating whether to parse Markdown tags.
This can be overridden locally by the tags '@md\code{and}@noMd` inside a
Roxygen comment.}

\item{file_encoding}{The file encoding. Default: \code{"UTF-8"}.}
}
\value{
A list of roxygen2 blocks.
}
\description{
\code{\link[=parse_code]{parse_code()}} takes a file or character vector, and an environment with R
values, and parses the roxygen2 comments found in it.
}
\keyword{internal}
