% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{parsers}
\alias{parse.default}
\alias{parse.name}
\alias{parse.name.description}
\alias{parse.toggle}
\alias{parse.unknown}
\alias{parse.value}
\alias{parse.words}
\alias{parse.words.line}
\alias{parsers}
\title{Parsers.}
\usage{
parse.default(key, rest)

parse.unknown(key, rest)

parse.value(key, rest)

parse.words(key, rest)

parse.words.line(key, rest)

parse.name.description(key, rest)

parse.name(key, name)

parse.toggle(key, rest)
}
\arguments{
  \item{key}{the parsing key}

  \item{rest}{the expression to be parsed}
}
\description{
These function implement parsing different tag types.
}
\details{
\code{parse.default}: emit value unchanged.

\code{parse.default}: throws errors about unknown tag.

\code{parse.value}: fail if empty

\code{parse.words}: parse values into words separated by space

\code{parse.words.one}: parse values into words separated by space
 (one line only)

\code{parse.description}: parse mandatory name and description

\code{parse.name}: one and only one word

\code{parse.toggle}: turn binary element on
}
\keyword{internal}

