\name{preorder.flatten.expression}
\alias{preorder.flatten.expression}
\title{Flatten a nested expression into a list, preorderly.}
\usage{preorder.flatten.expression(expression)}
\description{Flatten a nested expression into a list, preorderly.}
\value{A list containing the flattened expression}
\arguments{\item{expression}{the root of the expression to be
flattened}}
