\name{prerefs}
\alias{prerefs}
\title{Comment blocks (possibly null) that precede a file's expressions.}
\usage{prerefs(srcfile, srcrefs)}
\description{Comment blocks (possibly null) that precede a file's expressions.}
\value{A list of prerefs that resemble srcrefs in form, i.e.
with srcfile and lloc}
\arguments{\item{srcfile}{result of running \code{srcfile} on an interesting
file}
\item{srcrefs}{the resultant srcrefs}}
