\name{Reduce.paste}
\alias{Reduce.paste}
\title{Ad-hoc abstraction to paste processed list-elements together.}
\usage{Reduce.paste(proc, elts, sep)}
\description{Ad-hoc abstraction to paste processed list-elements together.}
\value{The processed elements as a glued string}
\arguments{\item{proc}{the procedure to apply to the elements}
\item{elts}{the elements to be processed}
\item{sep}{the glue to joined the processed elements}}
