% Generated by roxygen2 (4.0.0): do not edit by hand
\name{register.preref.parser}
\alias{register.preref.parser}
\alias{register.preref.parsers}
\title{Register parsers.}
\usage{
register.preref.parser(key, parser)

register.preref.parsers(parser, ...)
}
\arguments{
\item{key,...}{Tag name(s).}

\item{parser}{Parser callback to register, a function with arguments
\code{key} and \code{expression}.}
}
\value{
\code{NULL}
}
\description{
Register parsers.
}
\keyword{internal}

