\name{register.preref.parsers}
\alias{register.preref.parsers}
\title{Register many parsers at once.}
\usage{
  register.preref.parsers(parser, ...)
}
\arguments{
  \item{parser}{the parser to register}

  \item{\dots}{the keys upon which to register}
}
\value{
  \code{NULL}
}
\description{
  Register many parsers at once.
}
\keyword{internal}

