\name{register.preref.parser}
\alias{register.preref.parser}
\title{Specifically register a preref parser...}
\arguments{
  \item{key}{the key upon which to register}
  \item{parser}{the parser callback to register;
a function taking \code{key} and \code{expression}}
}
\description{
  Specifically register a preref parser
}
\seealso{\code{\link{register.parser}}}
\value{\code{NULL}}
\keyword{internal}

