\name{register.preref.parsers}
\alias{register.preref.parsers}
\title{Register many preref parsers at once.}
\arguments{
  \item{parser}{the parser to register}
  \item{\dots}{the keys upon which to register}
}
\description{
  Register many preref parsers at once.
}
\value{\code{NULL}}
\keyword{internal}

