% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-registry.R
\name{register_tags}
\alias{register_tags}
\title{Register tag parsers.}
\usage{
register_tags(...)
}
\arguments{
\item{...}{List of name-parser pairs}
}
\description{
Use this if you're building your own roclet that uses new tags.
}
\keyword{internal}

