\name{roc_proc}
\alias{roc_proc}
\title{Process roclet and capture results.}
\usage{roc_proc(roclet, paths)}
\arguments{
  \item{roclet}{to use for processing}
  \item{input}{character vector of paths to files to process}
}
\description{
  Process roclet and capture results.
}
\seealso{\code{\link{roxygenise}} for user-friendly interface}
\keyword{internal}

