% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roclet_find}
\alias{roclet_find}
\title{Create a roclet from a string}
\usage{
roclet_find(x)
}
\arguments{
\item{x}{Arbitrary R code evaluated in roxygen2 package.}
}
\description{
This provides a flexible way of specifying a roclet in a string.
}
\examples{
# rd, namespace, and vignette work for backward compatibility
roclet_find("rd")

# But generally you should specify the name of a function that
# returns a roclet
roclet_find("rd_roclet")

# If it lives in another package, you'll need to use ::
roclet_find("roxygen2::rd_roclet")

# If it takes parameters (which no roclet does currently), you'll need
# to call the function
roclet_find("roxygen2::rd_roclet()")
}
\seealso{
Other extending: 
\code{\link{is_s3_generic}()},
\code{\link{load_options}()},
\code{\link{object}()},
\code{\link{object_format}()},
\code{\link{parse_package}()},
\code{\link{rd_section}()},
\code{\link{roc_proc_text}()},
\code{\link{roxy_block}()},
\code{\link{roxy_tag}()},
\code{\link{roxy_tag_rd}()},
\code{\link{tag_parsers}},
\code{\link{tags_list}()}
}
\concept{extending}
