% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R, R/utils-warn.R
\name{roxy_tag}
\alias{roxy_tag}
\alias{roxy_tag_parse}
\alias{roxy_tag_warning}
\alias{warn_roxy_tag}
\title{\code{roxy_tag} S3 constructor}
\usage{
roxy_tag(tag, raw, val = NULL, file = NA_character_, line = NA_character_)

roxy_tag_parse(x)

roxy_tag_warning(x, ...)

warn_roxy_tag(tag, message, parent = NULL, envir = parent.frame())
}
\arguments{
\item{tag}{Tag name. Arguments starting with \code{.} are reserved for internal
usage.}

\item{raw}{Raw tag value, a string.}

\item{val}{Parsed tag value, typically a character vector, but sometimes
a list. Usually filled in by \code{tag_parsers}}

\item{file, line}{Location of the tag}

\item{x}{A tag}

\item{...}{Additional data to be stored in the condition object.
If you supply condition fields, you should usually provide a
\code{class} argument. You may consider prefixing condition fields
with the name of your package or organisation to prevent name
collisions.}

\item{message}{Warning message}

\item{parent}{Supply \code{parent} when you rethrow an error from a
condition handler (e.g. with \code{\link[rlang:try_fetch]{try_fetch()}}).
\itemize{
\item If \code{parent} is a condition object, a \emph{chained error} is
created, which is useful when you want to enhance an error with
more details, while still retaining the original information.
\item If \code{parent} is \code{NA}, it indicates an unchained rethrow, which
is useful when you want to take ownership over an error and
rethrow it with a custom message that better fits the
surrounding context.

Technically, supplying \code{NA} lets \code{abort()} know it is called
from a condition handler. This helps it create simpler
backtraces where the condition handling context is hidden by
default.
}

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-chaining]{Including contextual information with error chains}}{\link[rlang:topic-error-chaining]{Including contextual information with error chains}}.}

\item{envir}{passed to \code{rlang::warn()} as \code{.envir}.}
}
\description{
\code{roxy_tag()} is the constructor for tag objects.
 \code{roxy_tag_warning()} is superseded by \code{warn_roxy_tag()}; use to generate a
warning that includes the location of the tag.
}
\section{Methods}{

Define a method for \code{roxy_tag_parse} to support new tags. See \link{tag_parsers}
for more details.
}

\seealso{
Other extending:
\code{\link[=load_options]{load_options()}},
\code{\link[=parse_package]{parse_package()}},
\code{\link[=rd_section]{rd_section()}},
\code{\link[=roc_proc_text]{roc_proc_text()}},
\code{\link[=roclet_find]{roclet_find()}},
\code{\link[=roxy_block]{roxy_block()}},
\code{\link[=roxy_tag_rd]{roxy_tag_rd()}},
\code{\link{tag_parsers}},
\code{\link[=tags_list]{tags_list()}}
}
\concept{extending}
