\name{roxygen-package}
\alias{roxygen-package}
\title{Literate Programming in R}
\docType{package}
\description{
  Roxygen is a Doxygen-like documentation system for R;
  allowing in-source specification of Rd files, collation
  and namespace directives.
}
\details{\tabular{ll}{ Package: \tab Roxygen\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab
  2008-08-25\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }}
\author{Peter Danenberg \email{pcd@roxygen.org},
Manuel Eugster \email{Manuel.Eugster@stat.uni-muenchen.de}

Maintainer: Peter Danenberg \email{pcd@roxygen.org}}
\seealso{See \code{\link{namespace_roclet}},
\code{\link{collate_roclet}},
for an overview of roxygen tags.}
\keyword{package}
\examples{\dontrun{roxygenize('pkg')}}

