% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygenize.R
\name{roxygenize}
\alias{roxygenize}
\alias{roxygenise}
\title{Document a package with roxygen2}
\usage{
roxygenize(package.dir = ".", roclets = NULL, load_code = NULL, clean = FALSE)

roxygenise(package.dir = ".", roclets = NULL, load_code = NULL, clean = FALSE)
}
\arguments{
\item{package.dir}{Location of package top level directory. Default is
working directory.}

\item{roclets}{Character vector of \link[=roclet]{roclets} to use.

The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}. This will update
(if needed) the \code{Collate} field with \code{\link[=update_collate]{update_collate()}},
produce the \code{NAMESPACE} file with \code{\link[=namespace_roclet]{namespace_roclet()}}, and
produce the Rd files with \code{\link[=rd_roclet]{rd_roclet()}}.

(Note that \code{update_collate()} is not technically a roclet but is still
controlled with this argument for historical reasons.)}

\item{load_code}{A function used to load all the R code in the package
directory. The default, \code{NULL}, uses the strategy defined by
the \code{load} roxygen option, which defaults to \code{\link[=load_pkgload]{load_pkgload()}}.
 See \link{load} for more details.}

\item{clean}{If \code{TRUE}, roxygen will delete all files previously
created by roxygen before running each roclet.}
}
\value{
\code{NULL}
}
\description{
This is the workhorse function that builds manual pages and metadata for a
package. It is powered by \link[=roclet]{roclets}, roxygen2's plugin system for
producing different types of output. See the documentation of the
individual components (\code{\link[=rd_roclet]{rd_roclet()}}, \code{\link[=namespace_roclet]{namespace_roclet()}},
\code{\link[=update_collate]{update_collate()}}) for more details, or learn how to make your own in
\code{vignette("extending")}.
}
\details{
Note that roxygen2 is a dynamic documentation system: it works by
inspecting loaded objects in the package. This means that you must
be able to load the package in order to document it: see \link{load} for
details.
}
