% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_package}
\alias{source_package}
\title{Source all files in a package.}
\usage{
source_package(path)
}
\arguments{
\item{path}{Path to a package.}
}
\value{
An environment, into which all R files in the directory were
sourced.
}
\description{
This is a wrapper around \link[pkgload:load_all]{pkgload::load_all}
}
\keyword{internal}
