% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-index-crossref}
\alias{tags-index-crossref}
\alias{@aliases}
\alias{@backref}
\alias{@concept}
\alias{@family}
\alias{@keywords}
\alias{@references}
\alias{@seealso}
\title{Tags for indexing and cross-references}
\usage{
#' @aliases ${1:alias}
#' @backref ${1:path}
#' @concept ${1:concept}
#' @family ${1:family name}
#' @keywords ${1:keyword}
#' @references ${1:reference}
#' @seealso [${1:func}()]
}
\description{
Learn the full details in \code{vignette('index-crossref')}.

Key tags:
\itemize{
\item \verb{@aliases $\{1:alias\}}: Add additional aliases to the topic.
 Use \code{NULL} to suppress the default alias automatically generated by roxygen2.
\item \verb{@concept $\{1:concept\}}: Add additional keywords or phrases to be included in the \code{help.search()} index. Each \verb{@concept} should be a single term or phrase.
\item \verb{@family $\{1:family name\}}: Generate \verb{@seealso} entries to all other functions in \verb{family name}.
\item \verb{@keywords $\{1:keyword\}}: Add a standardised keyword, indexed by \code{help.search()}. These are generally not useful apart from \verb{@keywords internal} which flags the topic as internal and removes from topic indexes.
\item \verb{@references $\{1:reference\}}: Pointers to the related literature. Usually formatted like a bibliography.
\item \verb{@seealso [$\{1:func\}()]}: Link to other related functions or urls. Usually a sentence or two, or a bulleted list.
}

Other less frequently used tags:
\itemize{
\item \verb{@backref $\{1:path\}}: Manually override the backreference that points from the \code{.Rd} file back to the source \code{.R} file. Only needed when generating code.
}
}
\seealso{
Other documentation tags:
\code{\link{tags-rd-R6}},
\code{\link{tags-rd-S3}},
\code{\link{tags-rd-S4}},
\code{\link{tags-rd-S7}},
\code{\link{tags-rd-datasets}},
\code{\link{tags-rd-functions}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
