% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-formatting}
\alias{tags-rd-formatting}
\alias{@md}
\alias{@noMd}
\alias{@section}
\title{Tags related to markdown support}
\usage{
#' @md
#' @noMd
#' @section {section title}:
}
\description{
Learn full the details in in \code{vignette('rd-formatting')}.

Other less frequently used tags:
\itemize{
\item \verb{@md}: Force markdown processing for a block.
\item \verb{@noMd}: Suppress markdown processing for a block.
\item \verb{@section \{section title\}:}: Add an arbitrary section to the documentation. Now generally superseded in favour of using a level 1 heading.
}
}
