% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-other}
\alias{@field}
\alias{@format}
\alias{@method}
\alias{@slot}
\alias{@source}
\title{Tags for documenting other objects}
\usage{
#' @field {name} {description}
#' @format {description}
#' @method {generic} {class}
#' @slot {name} {description}
#' @source {description}
}
\description{
Tags are briefly described below. Learn more about their usage in \code{vignette('rd-other')}.

\itemize{
\item \verb{@field \{name\} \{description\}}: Describe a R6 or refClass field.
\item \verb{@format \{description\}}: Describe the type/shape of a dataset. If the dataset is a data frame, include a description of each column. If not supplied, will be automatically generated by \code{object_format()}
\item \verb{@method \{generic\} \{class\}}: Force a function to be recognised as an S3 method. This affects the default usage and the \code{NAMESPACE} directive produced by \verb{@export}. Only needed if automatic detection fails.
\item \verb{@slot \{name\} \{description\}}: Describe the slot of an S4 class.
\item \verb{@source \{description\}}: Describe where the dataset came from. If possible, link to the original source and briefly describe any manipulation that you performed.
}
}
