% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-R6}
\alias{tags-rd-R6}
\alias{@field}
\alias{@R6method}
\title{Tags for documenting R6}
\usage{
#' @field ${1:name} ${2:description}
#' @R6method ${1:Class}$${2:method}
}
\description{
Learn the full details in \code{vignette('rd-R6')}.

Key tags:
\itemize{
\item \verb{@field $\{1:name\} $\{2:description\}}: Describe a R6 or refClass field.
\item \verb{@R6method $\{1:Class\}$$\{2:method\}}: Document an R6 method that can't be discovered by introspection, such as methods added via \verb{$set()}.
}
}
\seealso{
Other documentation tags:
\code{\link{tags-index-crossref}},
\code{\link{tags-rd-S3}},
\code{\link{tags-rd-S4}},
\code{\link{tags-rd-S7}},
\code{\link{tags-rd-datasets}},
\code{\link{tags-rd-functions}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
