% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-S4}
\alias{tags-rd-S4}
\alias{@slot}
\title{Tags for documenting S4}
\usage{
#' @slot ${1:name} ${2:description}
}
\description{
Learn the full details in \code{vignette('rd-S4')}.

Key tags:
\itemize{
\item \verb{@slot $\{1:name\} $\{2:description\}}: Describe the slot of an S4 class.
}
}
\seealso{
Other documentation tags:
\code{\link{tags-index-crossref}},
\code{\link{tags-rd-R6}},
\code{\link{tags-rd-S3}},
\code{\link{tags-rd-S7}},
\code{\link{tags-rd-datasets}},
\code{\link{tags-rd-functions}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
