% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-S7}
\alias{tags-rd-S7}
\alias{@prop}
\title{Tags for documenting S7}
\usage{
#' @prop ${1:name} ${2:description}
}
\description{
Learn the full details in \code{vignette('rd-S7')}.

Key tags:
\itemize{
\item \verb{@prop $\{1:name\} $\{2:description\}}: Describe an S7 class property that is not a constructor parameter.
}
}
\seealso{
Other documentation tags:
\code{\link{tags-index-crossref}},
\code{\link{tags-rd-R6}},
\code{\link{tags-rd-S3}},
\code{\link{tags-rd-S4}},
\code{\link{tags-rd-datasets}},
\code{\link{tags-rd-functions}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
