% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd}
\alias{@backref}
\alias{@description}
\alias{@details}
\alias{@examplesIf}
\alias{@family}
\alias{@noRd}
\alias{@param}
\alias{@returns}
\alias{@seealso}
\title{Tags for documenting functions}
\usage{
#' @backref {path}
#' @description
#' @details
#' @examplesIf {condition}
#' @family {family name}
#' @noRd
#' @param {name} {description}
#' @returns {description}
#' @seealso {links to other functions}
}
\description{
Tags are briefly described below. Learn more about their usage in vignette('rd').

\itemize{
\item \verb{@backref \{path\}}: Manually override the backreference that points back from the \code{.Rd} to the source \code{.R} file. Only needed when generating code.
\item \verb{@description}: A short description of the purpose of the function. Usually a paragraph or less.
\item \verb{@details}: Additional details about the function. Now generally superseded by using a level 1 markdown heading.
\item \verb{@examplesIf \{condition\}}: Run examples only if a \code{condition} is \code{TRUE}.
\item \verb{@family \{family name\}}: Automatically generate \verb{@seealso} entries for all functions in family.
\item \verb{@noRd}: Suppress \code{.Rd} generation for a block. Use for documentation blocks that should only be visible in the source code.
\item \verb{@param \{name\} \{description\}}: Document a function input. Should describe acceptable input types and how it affects the output. \code{description} is usually one or two sentences but can be as long as needed. Document multiple arguments by separating names with commas (without spaces).
\item \verb{@returns \{description\}}: The type of output returned by the function.
\item \verb{@seealso \{links to other functions\}}: Other related functions.
}
}
