% Generated by roxygen2 (4.0.0): do not edit by hand
\name{upgradeRoxygen}
\alias{upgradeRoxygen}
\title{Run this before upgrading to 4.0.0.}
\usage{
upgradeRoxygen(path = ".")
}
\arguments{
  \item{path}{Package path to upgrade. Defaults to working
  directory.}
}
\description{
This function deletes all files in \file{man/} and flags \file{NAMESPACE}
as being safe to override. Run this once before you upgrade to roxygen2
4.0.0 and NEVER AGAIN.
}

