% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-usage.R
\name{usage-internals}
\alias{usage-internals}
\alias{roxygen_test_fun,function-method}
\alias{roxygen_test_fun,<--method}
\alias{mean.function}
\alias{mean.<-}
\title{Test usage generation}
\usage{
\S4method{roxygen_test_fun}{`function`}(x)

\S4method{roxygen_test_fun}{`<-`}(x)

\method{mean}{`function`}(x)

\method{mean}{`<-`}(x)
}
\description{
Test usage generation
}
\keyword{internal}
