\name{usage}
\alias{usage}
\title{Given argument list, produce usage string for it.}
\usage{usage(args)}

\description{
  Given argument list, produce usage string for it.
}

\details{
  Adapted from \code{\link{prompt}}.
}
\value{a string}
\arguments{
  \item{f}{function, or name of function, as string}
}
