% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-doxygen.R
\name{use_doxygen}
\alias{use_doxygen}
\title{Prepare the package so that it uses doxygen}
\usage{
use_doxygen(pkg = ".")
}
\arguments{
\item{pkg}{The package folder}
}
\description{
Calls doxygen_init with default directory and add 
     doxygen to the list of roclets to be used by default for this package
}
\note{
Please note this implies the installation of doxygen on the system,
      accessible on the command line by `doxygen`
}
\examples{
\dontrun{
use_doxygen()
}
}

