% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-escaping.R
\name{double_escape_md}
\alias{double_escape_md}
\title{Check markdown escaping}
\usage{
double_escape_md(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
Double-escaped text.
}
\description{
This is a regression test for Markdown escaping.
}
\details{
Each of the following bullets should look the same when rendered:
\itemize{
\item Double escapes: \\, \\\%, \\$, \\_
\item Backticks: \verb{\\}, \verb{\\\%}, \verb{\\$}, \verb{\\_}
\item \verb{\verb{}}: \verb{\\}, \verb{\\\%}, \verb{\$}, \verb{\_}
}

[ this isn't a link ]
\[ neither is this \]
}
\examples{
"\%" # percent
"\"" # double quote
'\'' # single quote
}
\keyword{internal}
