% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-escaping.R
\name{double_escape_md}
\alias{double_escape_md}
\title{Escape \% and \$ and \_ once more, because commonmark
removes the escaping. We do this everywhere currently.}
\usage{
double_escape_md(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
Double-escaped text.
}
\description{
Escape \% and \$ and \_ once more, because commonmark
removes the escaping. We do this everywhere currently.
}
