% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-examples.R
\name{escape_examples}
\alias{escape_examples}
\title{Escape examples}
\usage{
escape_examples(x)
}
\description{
This documentation topic is used primarily for testing and to record
our understanding of the \verb{\\example\{\}} escaping rules.
See \url{https://developer.r-project.org/parseRd.pdf} for the details provided
by R core.
}
\examples{
# In examples we automatically escape Rd comments (\%):
100 \%\% 30
# even if they are in strings
"50\%"

# and special backslashes inside of strings and symbols
"\\link{} \\var{}" # \link{}
"\\\\link{} \\\\var{}" # \\link{}
"\\\\"
`\\\\`

# (but only \\, \l, and \v need special treatment)
"\""
"\n"

# Otherwise, backslashes and parentheses are left as is.
# This means that you'll need to escape backslashes in infix operators
`\%\\\\\%` <- function(x, y) x + y
10 \%\\\\\% 20

# And unbalanced parentheses, which typically only occur in \dontshow{}:
\dontshow{if (FALSE) \{ }
print("Hello")
\dontshow{ \} }

# Otherwise, you _can_ escape parentheses, but there's little point.
# The following two lines are equivalent
f <- function() { NULL }
f <- function() \{ NULL \}
}
\keyword{internal}
