% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-examples.R
\name{escape_examples}
\alias{escape_examples}
\title{Escape examples}
\usage{
escape_examples(x)
}
\description{
Escape examples
}
\examples{
# Rd comments are escaped automatically
100 \%\% 30
"50\%"

# Braces are left as is
1 # \link{mean}
"{"

# As are backslashes
"\""

\dontshow{if (FALSE) \{ }
print("Hello")
\dontshow{ \} }
}
\keyword{internal}
