\name{expression.from.partitum}
\alias{expression.from.partitum}
\title{Extract the expression from the parse tree.}
\description{Extract the expression from the parse tree.}
\usage{expression.from.partitum(partitum)}
\arguments{\item{partitum}{partitum the parsed elements}}
\value{the extracted expression}
