% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{generate_template}
\alias{generate_template}
\title{Generate a template params list}
\usage{
generate_template(x)
}
\arguments{
\item{x}{name of the function for which you want a param template}
}
\description{
This function generates a template params list for roxygen
Add more functionality to get probably values, title, description etc
}

