% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygenize.R
\name{load_options}
\alias{load_options}
\title{Load options from DESCRIPTION.}
\usage{
load_options(base_path = ".")
}
\arguments{
\item{base_path}{Path to package.}
}
\description{
Load options from DESCRIPTION.
}
\keyword{internal}

