% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-link.R
\name{markdown-test}
\alias{markdown-test}
\title{Dummy page to test roxygen's markdown formatting}
\description{
Links are very tricky, so I'll put in some links here:
Link to a function: \code{\link[=roxygenize]{roxygenize()}}.
Link to an object: \link{roxygenize} (we just treat it like an object here.
}
\details{
Link to another package, function: \code{\link[devtools:document]{devtools::document()}}.
Link to another package, non-function: \link[devtools:document]{devtools::document}.

Link with link text: \link[=roxygenize]{this great function},
\code{\link[=roxygenize]{roxygenize}}, or \link[=roxygenize]{that great function}.

In another package: \link[devtools:document]{and this one}.

Can we link to another package, without specifying which one?
E.g. \code{\link[=str_replace]{str_replace()}}. Or to a non-function topic: \link{str_replace}.
This is the correct filename, so all is good.

This one is not the correct filename, so we'll need to remove it before
CRAN submission: \code{\link[=str_replace_all]{str_replace_all()}} and \link{str_replace_all}.

This is a table:\tabular{lr}{
   \strong{foo} \tab \strong{bar} \cr
   1 \tab 2 \cr
   100 \tab 200 \cr
}
}
\keyword{internal}
