% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown_pass1}
\alias{markdown_pass1}
\title{Expand the embedded inline code}
\usage{
markdown_pass1(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
Text with the inline code expanded. A character vector of the
same length as the input \code{text}.
}
\description{
For example this becomes two: 2.
Variables can be set and then reused, within the same
tag:
The value of \code{x} is 100.
}
\details{
We have access to the internal functions of the package, e.g.
since this is \emph{roxygen2}, we can refer to the internal \code{markdown}
function, and this is \code{TRUE}: TRUE.

Roxygen created this manual page using R version 3.6.1.

To insert the name of the current package: roxygen2.

The \code{iris} data set has 5 columns:
\code{Sepal.Length}, \code{Sepal.Width}, \code{Petal.Length}, \code{Petal.Width}, \code{Species}.
}
\keyword{internal}
