% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{markdown_pass1}
\alias{markdown_pass1}
\title{Expand R code}
\usage{
markdown_pass1(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
Text with R code expanded.
A character vector of the same length as the input \code{text}.
}
\description{
Expand R code
}
\details{
See \code{vignette("rd-formatting")}.
}
\keyword{internal}
