% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{namespace_roclet}
\alias{namespace_roclet}
\title{Roclet: make \code{NAMESPACE}}
\usage{
namespace_roclet()
}
\description{
This \link{roclet} automates the production of a \code{NAMESPACE} file, which controls
the functions imported and exported by your package, as described in
\href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}{Writing R extensions}.
It is run by default by \code{\link[=roxygenize]{roxygenize()}}.

The \code{NAMESPACE} is generated in two passes: the first generates only
import directives (because this can be computed without evaluating package
code), and the second generates everything (after the package has been
loaded).

See \code{vignette("namespace")} for details.
}
\examples{
# The most common namespace tag is @export, which declares that a function
# is part of the external interface of your package
#' @export
foofy <- function(x, y, z) {
}
# This results in the following line in `NAMESPACE`:
# export(foofy)

# You'll also often find global imports living in a file called
# R/{package}-package.R.
#' @importFrom magrittr \%>\%
#' @import rlang
NULL

# This results in the following lines in `NAMESPACE`:
# importFrom(magrittr,"\%>\%")
# import(rlang)

# There is a new experimental way to bulk-import a package:
#' @importAllFrom rlang

# This results in the following lines:
# importFrom(rlang,
#   "!!!",
#   "!!",
#   "\%&&\%",
#   ...
# )

# The exported objects are all explicitly imported one by one, which prevents
# load-time issues for the users of your package when an update creates an
# import conflict. The only time a conflict can arise is when you regenerate
# your namespace file, at which point you can manually resolve any conflicts by
# excluding a symbol with a `-` prefix:
#' @importAllFrom rlang -list2 -`:=`
}
\seealso{
\link{tags-namespace} for tags that generate \code{NAMESPACE} directives.
}
