% Generated by roxygen2 (4.0.0): do not edit by hand
\name{namespace_roclet}
\alias{S3method}
\alias{export}
\alias{exportClass}
\alias{exportMethod}
\alias{import}
\alias{importClassesFrom}
\alias{importFrom}
\alias{importMethodsFrom}
\alias{namespace_roclet}
\title{Roclet: make NAMESPACE.}
\usage{
namespace_roclet()
}
\description{
This roclet automates the production of a \file{NAMESPACE} file,
see \cite{Writing R Extensions}
(\url{http://cran.r-project.org/doc/manuals/R-exts.pdf}) for details.
}
\examples{
#' An example file, example.R, which imports
#' packages foo and bar
#' @import foo bar
NULL

#' An exportable function
#' @export
fun <- function() {}

roclet <- namespace_roclet()
\dontrun{roc_proc(roclet, "example.R")}
\dontrun{roc_out(roclet, "example.R", ".")}
}
\seealso{
Other roclets: \code{\link{rd_roclet}}
}

