% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{namespace_roclet}
\alias{namespace_roclet}
\alias{export}
\alias{exportClass}
\alias{exportMethod}
\alias{S3method}
\alias{import}
\alias{importFrom}
\alias{importClassesFrom}
\alias{importMethodsFrom}
\alias{rawNamespace}
\alias{useDynLib}
\title{Roclet: make NAMESPACE.}
\usage{
namespace_roclet()
}
\description{
This roclet automates the production of a \code{NAMESPACE} file,
see Writing R Extensions.
(\url{https://cran.r-project.org/doc/manuals/R-exts.pdf}) for details.
}
\seealso{
\code{vignette("namespace", package = "roxygen2")}

Other roclets: \code{\link{rd_roclet}},
  \code{\link{vignette_roclet}}
}
