% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{namespace_roclet}
\alias{namespace_roclet}
\alias{export}
\alias{exportClass}
\alias{exportMethod}
\alias{exportPattern}
\alias{import}
\alias{importFrom}
\alias{importClassesFrom}
\alias{importMethodsFrom}
\alias{evalNamespace}
\alias{rawNamespace}
\alias{S3method}
\alias{useDynLib}
\title{Roclet: make NAMESPACE.}
\usage{
namespace_roclet()
}
\description{
This roclet automates the production of a \code{NAMESPACE} file,
see Writing R Extensions.
(\url{https://cran.r-project.org/doc/manuals/R-exts.pdf}) for details.
The \code{NAMESPACE} is generated in two passes: the first generates only
import directives (because this can be computed without evaluating package
code), and the second generates everything (after the packaege has been
loaded).
}
\seealso{
\code{vignette("namespace", package = "roxygen2")}

Other roclets: \code{\link{rd_roclet}},
  \code{\link{vignette_roclet}}
}
\concept{roclets}
