% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-namespace.R
\name{namespace_roclet}
\alias{S3method}
\alias{export}
\alias{exportClass}
\alias{exportMethod}
\alias{import}
\alias{importClassesFrom}
\alias{importFrom}
\alias{importMethodsFrom}
\alias{namespace_roclet}
\title{Roclet: make NAMESPACE.}
\usage{
namespace_roclet()
}
\description{
This roclet automates the production of a \file{NAMESPACE} file,
see \cite{Writing R Extensions}
(\url{http://cran.r-project.org/doc/manuals/R-exts.pdf}) for details.
}
\seealso{
\code{vignette("namespace", package = "roxygen2")}

Other roclets: \code{\link{rd_roclet}},
  \code{\link{vignette_roclet}}
}

