% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygenize-needs.R
\name{needs_roxygenize}
\alias{needs_roxygenize}
\title{Check if documentation needs to be updated}
\usage{
needs_roxygenize(package.dir = ".")
}
\arguments{
\item{package.dir}{Location of package top level directory. Default is
working directory.}
}
\value{
A logical value, invisibly. \code{TRUE} if any man pages appear
to be out of date; \code{FALSE} otherwise.
}
\description{
A lightweight check that compares modification times of \code{.Rd} files
in \verb{man/} with the source files listed in their backrefs. This is
much faster than running \code{\link[=roxygenize]{roxygenize()}} but may report false
positives, e.g. if you've changed a roxygen template or an
inherited documentation topic.
}
\examples{
\dontrun{
needs_roxygenize()
}
}
