% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object}
\alias{object}
\title{Constructors for S3 object to represent R objects.}
\usage{
object(value, alias = NULL, type = obj_type(value))
}
\arguments{
\item{value}{The object itself.}

\item{alias}{Alias for object being documented, in case you create a
generator function with different name.}
}
\description{
These objects are usually created by the parsers, but it is also
useful to generate them by hand for testing.
}
\keyword{internal}

