\name{parse.file}
\alias{parse.file}
\title{Parse a source file containing roxygen directives.}
\usage{parse.file(file)}
\arguments{
  \item{file}{string naming file to be parsed}
}
\description{
  Parse a source file containing roxygen directives.
}
\value{List containing parsed directives}
\keyword{internal}

