\name{parse.file}
\alias{parse.file}
\title{Parse a source file containing roxygen directives.}
\usage{
  parse.file(file, env, env_hash)
}
\arguments{
  \item{file}{string naming file to be parsed}
}
\value{
  List containing parsed directives
}
\description{
  Parse a source file containing roxygen directives.
}
\keyword{internal}

