\name{parse.file}
\alias{parse.file}
\title{Parse a source file containing roxygen directives.}
\usage{### parse.file(file)
parse.file(...)
}
\description{Parse a source file containing roxygen directives.}
\value{List containing parsed directives}
\arguments{\item{...}{string naming file to be parsed}}
