\name{parse.name}
\alias{parse.name}
\title{Parse an element containing a single name and only a name;...}
\usage{
  parse.name(key, name)
}
\arguments{
  \item{key}{parsing key}

  \item{name}{the name to be parsed}
}
\value{
  A list containing key and name
}
\description{
  Parse an element containing a single name and only a
  name; extra material will be ignored and a warning
  issued.
}
\keyword{internal}

