\name{parse.name}
\alias{parse.name}
\title{Parse an element containing a single name and only a name;...}
\usage{parse.name(key, name)}
\description{Parse an element containing a single name and only a name;
extra material will be ignored and a warning issued.}
\value{A list containing key and name}
\arguments{\item{key}{parsing key}
\item{name}{the name to be parsed}}
