% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_roclet}
\alias{rd_roclet}
\title{Roclet: make Rd files}
\usage{
rd_roclet()
}
\description{
This \link{roclet} automates the production of the \code{.Rd} files that
R uses to document functions, datasets, packages, classes, and more.
See \code{vignette("rd")} for details.

It is run by default by \code{\link[=roxygenize]{roxygenize()}}.
}
\examples{
#' Add together two numbers
#' @param x A number.
#' @param y A number.
#' @return A number.
#' @export
#' @examples
#' add(1, 1)
#' add(10, 1)
add <- function(x, y) {
  x + y
}
}
\seealso{
\link{tags-rd}, \link{tags-rd-other}, \link{tags-reuse}, \link{tags-index-crossref} for
tags provided by this roclet.
}
