% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rd_roclet}
\alias{rd_roclet}
\title{Roclet: make Rd files.}
\usage{
rd_roclet()
}
\description{
This roclet is the workhorse of \pkg{roxygen}, producing the Rd files that
document that functions in your package.
}
\details{
This roclet also automatically runs \code{\link[tools]{checkRd}} on all
generated Rd files so that you know as early as possible if there's a
problem.
}
\examples{
roclet <- rd_roclet()
\dontrun{roc_proc(roclet, "example.R")}
\dontrun{roc_out(roclet, "example.R", ".")}
}
\seealso{
Other roclets: \code{\link{S3method}},
  \code{\link{export}}, \code{\link{exportClass}},
  \code{\link{exportMethod}}, \code{\link{import}},
  \code{\link{importClassesFrom}},
  \code{\link{importFrom}},
  \code{\link{importMethodsFrom}},
  \code{\link{namespace_roclet}}
}

