% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_roclet}
\alias{rd_roclet}
\title{Roclet: make Rd files}
\usage{
rd_roclet()
}
\description{
This roclet is the workhorse of roxygen2, producing the \code{.Rd} files that
R uses to document functions, datasets, packages, classes, and more.
See \code{vignette("rd")} for details.

Generally you will not call this function directly
but will instead use \code{\link[=roxygenise]{roxygenise()}} specifying the rd roclet.
}
\examples{
#' The length of a string (in characters)
#'
#' @param x A character vector.
#' @returns An integer vector the same length as `x`.
#'   `NA` strings have `NA` length.
#' @seealso [nchar()]
#' @export
#' @examples
#' str_length(letters)
#' str_length(c("i", "like", "programming", NA))
str_length <- function(x) {
}
}
\seealso{
\link{tags-rd}, \link{tags-rd-other}, \link{tags-reuse}, \link{tags-index-crossref} for
tags provided by this roclet.
}
