% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_roclet}
\alias{rd_roclet}
\title{Roclet: make Rd files.}
\usage{
rd_roclet()
}
\description{
This roclet is the workhorse of \code{roxygen}, producing the Rd files that
document that functions in your package.

Generally you will not call this function directly
but will instead use roxygenise() specifying the rd roclet
}
\seealso{
\code{vignette("rd", package = "roxygen2")}

Other roclets: \code{\link{namespace_roclet}},
  \code{\link{vignette_roclet}}
}
\concept{roclets}
