% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-rd.R
\name{rd_roclet}
\alias{rd_roclet}
\title{Roclet: make Rd files.}
\usage{
rd_roclet()
}
\description{
This roclet is the workhorse of \pkg{roxygen}, producing the Rd files that
document that functions in your package.
}
\seealso{
\code{vignette("rd", package = "roxygen2")}

Other roclets: \code{\link{namespace_roclet}},
  \code{\link{vignette_roclet}}
}

