% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-registry.R
\name{register.preref.parser}
\alias{register.preref.parser}
\alias{register.preref.parsers}
\alias{register_tags}
\title{Register tag parsers.}
\usage{
register_tags(...)
}
\arguments{
\item{...}{List of name-parser pairs}
}
\value{
\code{NULL}
}
\description{
Register tag parsers.
}
\keyword{internal}

