\name{register.parsers}
\alias{register.parsers}
\title{Register many parsers at once.}
\usage{register.parsers(table, parser, ...)}
\description{Register many parsers at once.}
\value{\code{NULL}}
\arguments{\item{table}{the table under which to register}
\item{parser}{the parser to register}
\item{\dots}{the keys upon which to register}}
