\name{register.srcref.parser}
\alias{register.srcref.parser}
\title{Specifically register a srcref parser...}
\description{Specifically register a srcref parser}
\value{\code{NULL}}
\seealso{\code{\link{register.parser}}}
\arguments{\item{key}{the key upon which to register}
\item{parser}{the parser callback to register;
a function taking \code{key} and \code{expression}}}
