\name{register.srcref.parsers}
\alias{register.srcref.parsers}
\title{Register many srcref parsers at once.}
\arguments{
  \item{parser}{the parser to register}
  \item{\dots}{the keys upon which to register}
}
\description{
  Register many srcref parsers at once.
}
\value{\code{NULL}}
\keyword{internal}

