\name{roc_out}
\alias{roc_out}
\title{Process roclet and output results.}
\usage{
  roc_out(roclet, input, base_path)
}
\arguments{
  \item{roclet}{to use for processing}

  \item{input}{character vector of paths to files to
  process}

  \item{base_path}{base directory in which to save output}
}
\description{
  Process roclet and output results.
}
\seealso{
  \code{\link{roxygenise}} for user-friendly interface
}
\keyword{internal}

