% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roc_proc_text}
\alias{roc_proc_text}
\title{Process roclet on string and capture results}
\usage{
roc_proc_text(roclet, input, wd = NULL)
}
\arguments{
\item{roclet}{Name of roclet to use for processing.}

\item{input}{Source string}

\item{wd}{Working directory}
}
\description{
Useful for testing.
}
\seealso{
Other extending: 
\code{\link{is_s3_generic}()},
\code{\link{load_options}()},
\code{\link{object}()},
\code{\link{object_format}()},
\code{\link{parse_package}()},
\code{\link{rd_section}()},
\code{\link{roclet_find}()},
\code{\link{roxy_block}()},
\code{\link{roxy_tag}()},
\code{\link{roxy_tag_rd}()},
\code{\link{tag_parsers}},
\code{\link{tags_list}()}
}
\concept{extending}
