\name{roc_proc_text}
\alias{roc_proc_text}
\title{Process roclet on string and capture results.
Useful for testing.}
\usage{
roc_proc_text(roclet, input, options = list())
}
\arguments{
  \item{roclet}{to use for processing}

  \item{input}{source string}

  \item{options}{a list of options to control roxygen
  behaviour.  Currently only \code{wrap} is recognised.}
}
\description{
Process roclet on string and capture results.
Useful for testing.
}
\keyword{internal}

