\name{roc_proc_text}
\alias{roc_proc_text}
\title{Process roclet on string and capture results.
Useful for testing.}
\usage{
  roc_proc_text(roclet, input)
}
\arguments{
  \item{roclet}{to use for processing}

  \item{input}{source string}
}
\description{
  Process roclet on string and capture results. Useful for
  testing.
}
\keyword{internal}

