% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roc_proc_text}
\alias{roc_proc_text}
\title{Process roclet on string and capture results.}
\usage{
roc_proc_text(roclet, input, options = list())
}
\arguments{
\item{roclet}{Name of roclet to use for processing.}

\item{input}{Source string}

\item{options}{A list of options to control roxygen behaviour.
Currently only \code{wrap} is recognised.}
}
\description{
Useful for testing.
}
\keyword{internal}

