% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roclet}
\alias{roclet}
\alias{roclet}
\alias{roclet_output}
\alias{roclet_tags}
\alias{roclet_process}
\alias{roclet_clean}
\title{Build a new roclet.}
\usage{
roclet(subclass, ...)

roclet_output(x, results, base_path, ...)

roclet_tags(x)

roclet_process(x, parsed, base_path, global_options = list())

roclet_clean(x, base_path)
}
\description{
To create a new roclet, you will need to create a constructor function
that wraps \code{roclet}, and then implement methods for
\code{roclet_tags}, \code{roclet_process}, \code{roclet_output}, and \code{roclet_clean}.
}
\keyword{internal}
