% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roclet_find}
\alias{roclet_find}
\title{Create a roclet from a string.}
\usage{
roclet_find(x)
}
\arguments{
\item{x}{Arbitrary R code evaluated in roxygen2 package.}
}
\description{
This provides a flexible way of specifying a roclet in a string.
}
\examples{
# rd, namespace, and vignette work for backward compatibility
roclet_find("rd")

# But generally you should specify the name of a function that
# returns a roclet
roclet_find("rd_roclet")

# If it lives in another package, you'll need to use ::
roclet_find("roxygen2::rd_roclet")

# If it takes parameters (which no roclet does currently), you'll need
# to call the function
roclet_find("roxygen2::rd_roclet()")
}
\keyword{internal}
