% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{roxy_tag}
\alias{roxy_tag}
\alias{roxy_tag_warning}
\alias{tag_code}
\alias{tag_examples}
\alias{tag_markdown}
\alias{tag_markdown_restricted}
\alias{tag_name}
\alias{tag_name_description}
\alias{tag_toggle}
\alias{tag_value}
\alias{tag_words}
\alias{tag_words_line}
\title{Parsing tags.}
\usage{
roxy_tag(tag, val, file = "", line = 0)

roxy_tag_warning(x, ...)

tag_value(x)

tag_name(x)

tag_name_description(x)

tag_words(min = 0, max = Inf)

tag_words_line(x)

tag_toggle(x)

tag_code(x)

tag_examples(x)

tag_markdown(x)

tag_markdown_restricted(x)
}
\arguments{
\item{tag}{Tag name}

\item{val}{Tag value. When read from the file, this will be a string,
but after parsing can be a more complicated structure (typically
a character vector, but sometimes a list).}

\item{file, line}{Location of the tag}
}
\description{
\code{roxy_tag} constructs a tag object, and \code{roxy_tag_warning} makes
an informative warning using the location information stored in the tag.
The remainder of the tag functions parse the tag value, convert a string
into a richer R object, or providing informative warnings and returning
valid if the value is invalid.
}
\details{
One exception to the rule is \code{tag_words}: you call it with
arguments giving the \code{min} and \code{max} words, and it will
generate a tag parser for you.
}
\keyword{internal}

