% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{roxy_tag}
\alias{roxy_tag}
\alias{roxy_tag_parse}
\alias{roxy_tag_warning}
\title{\code{roxy_tag} S3 constructor}
\usage{
roxy_tag(tag, raw, val = NULL, file = NA_character_, line = NA_integer_)

roxy_tag_parse(x)

roxy_tag_warning(x, ...)
}
\arguments{
\item{tag}{Tag name. Arguments starting with \code{.} are reserved for internal
usage.}

\item{raw}{Raw tag value, a string.}

\item{val}{Parsed tag value, typically a character vector, but sometimes
a list. Usually filled in by \code{tag_parsers}}

\item{file, line}{Location of the tag}

\item{x}{A tag}
}
\description{
\code{roxy_tag()} is the constructor for tag objects.
\code{roxy_tag_warning()} generates a warning that gives the location of the tag.
}
\section{Methods}{

Define a method for \code{roxy_tag_parse} to support new tags. See \link{tag_parsers}
for more details.
}

\keyword{internal}
