% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{roxy_tag_rd}
\alias{roxy_tag_rd}
\title{Generate Rd output from a tag}
\usage{
roxy_tag_rd(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
Methods must return a \link{rd_section}.
}
\description{
Provide a method for this generic if you want a tag to generate output
in \code{.Rd} files. See \code{vignette("extending")} for more details.
}
\keyword{internal}
