% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{roxygen}
\alias{roxygen}
\alias{roxygen-package}
\title{In-line documentation for R.}
\description{
Roxygen is a Doxygen-like documentation system for R; allowing
in-source specification of Rd files, collation and namespace
directives.
}
\details{
If you have existing Rd files, check out the \code{Rd2roxygen} package
for a convenient way of converting Rd files to roxygen comments.
}
\examples{
\dontrun{roxygenize('pkg')}
}
\author{
Hadley Wickham \email{h.wickham@gmail.com},
Peter Danenberg \email{pcd@roxygen.org},
Manuel Eugster \email{Manuel.Eugster@stat.uni-muenchen.de}

Maintainer: Hadley Wickham \email{h.wickham@gmail.com}
}
\seealso{
See \code{vignette("roxygen2", package = "roxygen2")} for an overview
  of the package, \code{vignette("rd", package = "roxygen2")} for generating
  documentation, and \code{vignette("namespace", package = "roxygen2")} for
  generating the namespace specification.
}

