% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{roxygen2-package}
\alias{roxygen2}
\alias{roxygen2-package}
\title{In-Source Documentation for R}
\description{
A 'Doxygen'-like in-source documentation system
for Rd, collation, and 'NAMESPACE' files.
}
\details{
If you have existing Rd files, check out the \code{Rd2roxygen} package
for a convenient way of converting Rd files to roxygen comments.
}
\examples{
\dontrun{roxygenize('pkg')}
}
\author{
Hadley Wickham \email{h.wickham@gmail.com},
Peter Danenberg \email{pcd@roxygen.org},
Manuel Eugster \email{Manuel.Eugster@stat.uni-muenchen.de}

Maintainer: Hadley Wickham \email{h.wickham@gmail.com}
}
\seealso{
See \code{vignette("roxygen2", package = "roxygen2")} for an overview
  of the package, \code{vignette("rd", package = "roxygen2")} for generating
  documentation, and \code{vignette("namespace", package = "roxygen2")} for
  generating the namespace specification.
}

