% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_package}
\alias{source_package}
\title{Source all files in a package.}
\usage{
source_package(path)
}
\arguments{
\item{path}{Path to a package.}
}
\value{
An environment, into which all R files in the directory were
  sourced.
}
\description{
This is a simple attempt to load code in a package used by
\code{\link{roxygenize}}. It will work with simple packages, but fail if
there are compiled files, data files, etc. In that case, it's better to
use \code{\link[devtools]{document}()}.
}
\keyword{internal}

