% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-index-crossref}
\alias{@aliases}
\alias{@backref}
\alias{@concept}
\alias{@family}
\alias{@keywords}
\alias{@references}
\alias{@seealso}
\title{Tags for indexing and cross-references}
\usage{
#' @aliases {alias}
#' @backref {path}
#' @concept {concept}
#' @family {family name}
#' @keywords {keyword}
#' @references {reference}
#' @seealso {[func1()], <url1>}
}
\description{
Tags are briefly described below. Learn more about their usage in \code{vignette('index-crossref')}.

\itemize{
\item \verb{@aliases \{alias\}}: Add additional aliases to the topic. Use \code{NULL} to suppress the default alias automatically generated by roxygen2.
\item \verb{@backref \{path\}}: Manually override the backreference that points from the \code{.Rd} file back to the source \code{.R} file. Only needed when generating code.
\item \verb{@concept \{concept\}}: Add additional keywords or phrases to be included in the \code{help.search()} index. Each \verb{@concept} should be a single term or phrase.
\item \verb{@family \{family name\}}: Automatically generate \verb{@seealso} entries to all other functions in \verb{family name}.
\item \verb{@keywords \{keyword\}}: Add keywords, indexed by \code{help.search()}. These are generally not useful apart from \verb{@keyword internal} which flags the topic as internal and removes from topic indexes.
\item \verb{@references \{reference\}}: Pointers to the related literature. Usually formatted like a bibliography.
\item \verb{@seealso \{[func1()], <url1>\}}: Link to other related functions or urls.
}
}
