% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-other}
\alias{tags-rd-other}
\alias{@field}
\alias{@format}
\alias{@method}
\alias{@slot}
\alias{@source}
\title{Tags for documenting datasets and classes}
\usage{
#' @field ${1:name} ${2:description}
#' @format ${1:description}
#' @method ${1:generic} ${2:class}
#' @slot ${1:name} ${2:description}
#' @source ${1:description}
}
\description{
Learn full the details in in \code{vignette('rd-other')}.

Key tags:
\itemize{
\item \verb{@field $\{1:name\} $\{2:description\}}: Describe a R6 or refClass field.
\item \verb{@format $\{1:description\}}: Describe the type/shape of a dataset. If the dataset is a data frame, include a description of each column. If not supplied, will be automatically generated by \code{object_format()}.
\item \verb{@method $\{1:generic\} $\{2:class\}}: Force a function to be recognised as an S3 method. This affects the default usage and the \code{NAMESPACE} directive produced by \verb{@export}. Only needed if automatic detection fails.
\item \verb{@slot $\{1:name\} $\{2:description\}}: Describe the slot of an S4 class.
\item \verb{@source $\{1:description\}}: Describe where the dataset came from. Provide a link to the original source (if possible) and briefly describe any manipulation that you performed when importing the data.
}
}
\seealso{
Other documentation tags: 
\code{\link{tags-index-crossref}},
\code{\link{tags-rd}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
