% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd-S3}
\alias{tags-rd-S3}
\alias{@method}
\title{Tags for documenting S3}
\usage{
#' @method ${1:generic} ${2:class}
}
\description{
Learn the full details in \code{vignette('rd-S3')}.

Key tags:
\itemize{
\item \verb{@method $\{1:generic\} $\{2:class\}}: Force a function to be recognised as an S3 method. This affects the default usage and the \code{NAMESPACE} directive produced by \verb{@export}. Only needed if automatic detection fails.
}
}
\seealso{
Other documentation tags:
\code{\link{tags-index-crossref}},
\code{\link{tags-rd-R6}},
\code{\link{tags-rd-S4}},
\code{\link{tags-rd-S7}},
\code{\link{tags-rd-datasets}},
\code{\link{tags-rd-functions}},
\code{\link{tags-reuse}}
}
\concept{documentation tags}
