% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags-rd}
\alias{@description}
\alias{@details}
\alias{@example}
\alias{@examples}
\alias{@examplesIf}
\alias{@noRd}
\alias{@param}
\alias{@rawRd}
\alias{@returns}
\alias{@title}
\alias{@usage}
\title{Tags for documenting functions}
\usage{
#' @description
#' @details
#' @example {path}
#' @examples
#' @examplesIf {condition}
#' @noRd
#' @param {name} {description}
#' @rawRd {rd}
#' @returns {description}
#' @title {title}
#' @usage {function_call(arg1, arg2 = default, ...)}
}
\description{
Tags are briefly described below. Learn more about their usage in \code{vignette('rd')}.

\itemize{
\item \verb{@description}: A short description of the purpose of the function. Usually a paragraph or less.
\item \verb{@details}: Additional details about the function. Now generally superseded by using a level 1 markdown heading.
\item \verb{@example \{path\}}: Path to file containing R code to use as examples.
\item \verb{@examples}: Examples demonstrating how the function works.
\item \verb{@examplesIf \{condition\}}: Run examples only if a \code{condition} is \code{TRUE}.
\item \verb{@noRd}: Suppress \code{.Rd} generation for a block. Use for documentation blocks that should only be visible in the source code.
\item \verb{@param \{name\} \{description\}}: Document a function input. Should describe acceptable input types and how it affects the output. \code{description} is usually one or two sentences but can be as long as needed. Document multiple arguments by separating names with commas (without spaces).
\item \verb{@rawRd \{rd\}}: Insert literal text directly into the \code{.Rd} file.
\item \verb{@returns \{description\}}: The type of output returned by the function.
\item \verb{@title \{title\}}: A one-line description of the function shown in various indexes. An explicit \verb{@title} is not usually needed as the first paragraph in the roxygen block will be used as the title.
\item \verb{@usage \{function_call(arg1, arg2 = default, ...)\}}: Override the default usage generated by roxygen2. Only needed when roxygen2 fails to correctly derive the usage of your function.
}
}
