% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-metadata.R
\name{tags_list}
\alias{tags_list}
\alias{tags_metadata}
\title{Access metadata about built-in tags}
\usage{
tags_list(built_in = TRUE)

tags_metadata()
}
\description{
Access metadata about built-in tags
}
\keyword{internal}
