% Generated by roxygen2 (4.0.0): do not edit by hand
\name{upgradeRoxygen}
\alias{upgradeRoxygen}
\title{Run this before upgrading to 4.0.0.}
\usage{
upgradeRoxygen(path = ".")
}
\arguments{
\item{path}{Package path to upgrade. Defaults to working directory.}
}
\description{
This function flags all files in \file{man/} and \file{NAMESPACE}
as being created by roxygen2. It is safe to run this function multiple
times, but you shouldn't need to.
}

